+ Lose number conversion warning fix

+++ ShufflerTumblr/ShufflerTumblr/Classes/BlogPlaylistViewController.m
@@ -113,7 +113,7 @@
 
 #pragma UITableView delegates
 -(void) tableView:(UITableView *) tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
-    chosenRow = indexPath.row;
+    chosenRow = (int)indexPath.row;
     [tableView deselectRowAtIndexPath:indexPath animated: YES];
     
     [[Player sharedInstance] clear];
