+++ ShufflerTumblr/ShufflerTumblr/Classes/DiscoverViewController.m
@@ -117,7 +117,7 @@
 
 #pragma UITableView delegates
 -(void) tableView:(UITableView *) tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
-    chosenRow = indexPath.row;
+    chosenRow = (int)indexPath.row;
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
     
     UIStoryboard *storyboard = [UIStoryboard storyboardWithName:@"MainStoryboard" bundle:nil];
