+++ ShufflerTumblr/ShufflerTumblr/Classes/MyBlogsViewController.m
@@ -105,7 +105,7 @@
 
 #pragma UITableView Delegates
 -(void) tableView:(UITableView *) tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
-    chosenRow = indexPath.row;
+    chosenRow = (int)indexPath.row;
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
     
     BlogInfo *selectedBlogInfo = [tableObjects objectAtIndex:chosenRow];
